+++ src/fizzbuzz/model.py
@@ -1,0 +1,12 @@
+class Input:
+    def __init__(self, value):
+        self.value = value
+
+    def is_multiple_of_three(self):
+        return self.value % 3 == 0
+
+    def is_multiple_of_five(self):
+        return self.value % 5 == 0
+
+    def is_multiple_of_three_and_five(self):
+        return (self.value % 3 == 0) & (self.value % 5 == 0)
