+++ tests/fizzbuzz/test_app.py
@@ -2,32 +2,7 @@ import unittest
 
 from parameterized import parameterized
 
-class Input:
-    def __init__(self, value):
-        self.value = value
-
-    def is_multiple_of_three(self):
-        return self.value % 3 == 0
-
-    def is_multiple_of_five(self):
-        return self.value % 5 == 0
-
-    def is_multiple_of_three_and_five(self):
-        return (self.value % 3 == 0) & (self.value % 5 == 0)
-
-
-class FizzBuzzApp:
-
-    def run(self, value:int):
-        my_input = Input(value)
-        if my_input.is_multiple_of_three_and_five():
-            return "FizzBuzz"
-        elif my_input.is_multiple_of_three():
-            return "Fizz"
-        elif my_input.is_multiple_of_five():
-            return "Buzz"
-        return value
-
+from fizzbuzz.app import FizzBuzzApp
 
 
 class TestFizzBuzzApp(unittest.TestCase):
